+++ modules/governor/src/commands/execute.ts
@@ -1,6 +1,10 @@
 import { defineCommand, encodeAction } from "@evmcrispr/sdk";
 import type Governor from "..";
-import { collectBlockActions, hashDescription } from "../utils";
+import {
+  collectBlockActions,
+  governorExecutor,
+  hashDescription,
+} from "../utils";
 
 export default defineCommand<Governor>({
   name: "execute",
@@ -20,9 +24,12 @@ export default defineCommand<Governor>({
       description: "Block of commands making up the proposal",
     },
   ],
-  async run(_module, { governor, description, actions }, { interpreters }) {
+  async run(module, { governor, description, actions }, { interpreters }) {
     const { targets, values, calldatas, totalValue } =
-      await collectBlockActions("execute", actions, interpreters);
+      await collectBlockActions("execute", actions, interpreters, {
+        module,
+        sender: await governorExecutor(module, governor),
+      });
 
     const action = encodeAction(
       governor,